requirements.txt
Django==5.0.1
Pillow==9.5.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0 